+++ example/articles/forms.py
@@ -1,9 +1,10 @@
 from __future__ import (absolute_import, unicode_literals)
 
+from arctic.forms import SimpleSearchForm, QuickFiltersFormMixin
 from django import forms
+from django.db.models import Q
 
 from .models import Article
-from django.db.models import Q
 
 
 class ArticleForm(forms.ModelForm):
@@ -22,8 +23,36 @@ class AdvancedArticleSearchForm(forms.Form):
                                   required=False,
                                   label='Description')
 
+    def __init__(self, data):
+        # Reset data, but store for get_search_filter
+        self.stored_data = data
+        super(AdvancedArticleSearchForm, self).__init__(data)
+
     def get_search_filter(self):
         value = self.cleaned_data.get('description')
         if value:
             return Q(description__icontains=value)
         return Q()
+
+
+class FiltersAndSearchForm(QuickFiltersFormMixin, SimpleSearchForm):
+    filters_query_name = 'my_filters'
+
+    FILTER_BUTTONS = (
+        ('published', 'Is published'),
+        ('rabbit', 'Find rabbit')
+    )
+
+    def get_search_filter(self):
+        q = super(FiltersAndSearchForm, self).get_search_filter()
+
+        values = self.cleaned_data.get(self.filters_query_name)
+        conditions = {
+            'published': Q(published=True),
+            'rabbit': Q(description__icontains='rabbit')
+        }
+
+        for value in values:
+            q |= conditions.get(value, Q())
+
+        return q