+++ example/articles/views.py
@@ -8,7 +8,8 @@ from arctic.generics import (CreateView, DeleteView, ListView, TemplateView,
                              UpdateView)
 from collections import OrderedDict
 
-from .forms import ArticleForm, AdvancedArticleSearchForm
+from .forms import (ArticleForm, AdvancedArticleSearchForm,
+                    FiltersAndSearchForm)
 from .inlines import TagsInline
 from .models import (Article, Category, Tag)
 
@@ -27,8 +28,9 @@ class ArticleListView(ListView):
     paginate_by = 10
     model = Article
     fields = ['title', 'description', 'published', 'category', 'tags']
-    ordering_fields = ['title', 'description', 'published']
+    ordering_fields = ['title', 'description', 'category', 'published']
     search_fields = ['title']
+    simple_search_form_class = FiltersAndSearchForm
     advanced_search_form_class = AdvancedArticleSearchForm
     breadcrumbs = (('Home', 'index'), ('Article List', None))
     action_links = [
@@ -52,6 +54,9 @@ class ArticleListView(ListView):
     def get_published_field(self, row_instance):
         symbol = 'fa-check' if row_instance.published else 'fa-minus'
         return mark_safe('<i class="fa {}"></i>'.format(symbol))
+
+    def get_category_ordering_field(self):
+        return 'category__name'
 
     def get_field_actions(self, row_instance):
         action_links = [